+++ linopy/io.py
@@ -591,6 +591,8 @@ def to_file(
     """
     Write out a model to a lp or mps file.
     """
+    m._check_sos_unmasked()
+
     if fn is None:
         fn = Path(m.get_problem_file())
     if isinstance(fn, str):